+++ resources/views/pages/advisor.blade.php
@@ -1,0 +1,3 @@
+<x-layouts::app :title="__('Advisor Q&A')">
+    <livewire:advisor.ask />
+</x-layouts::app>